// frontend/src/types/IResult.ts
export interface IResult {
    result_id?: number;
    round_id?: number;
    competition_event_competition_athlete_id?: number;
    competition_event_competition_athlete?: number;
    result_as_number?: number;
    round?: number;
    lane_or_order_number?: number;
    wind_as_number?: number;
    attempt_nr?: number;
    points_from_result?: number;
    result_as_char?: string;
    round_number?: number;
    starting_height?: number;
}